+++ desktop/src/com/mygdx/game/desktop/View.java
@@ -20,7 +20,7 @@ public class View extends OrthogonalTiledMapRenderer {
     private static View single_instance = null;
     private ArrayList<Sprite> sprites;
     private int drawSpritesAfterLayer = 3;
-    private int scale;
+    private float scale;
 
     public static View getInstance() {
         if (single_instance == null)
@@ -28,13 +28,13 @@ public class View extends OrthogonalTiledMapRenderer {
         return single_instance;
     }
 
-    public static View createInstance(int scale) {
+    public static View createInstance(float scale) {
         if (single_instance == null)
             single_instance = new View(new TmxMapLoader().load("outside.tmx"), scale);
         return single_instance;
     }
 
-    public View(TiledMap map, int scale) {
+    public View(TiledMap map, float scale) {
         super(map, scale);
         sprites = new ArrayList<>();
     }
